requirements.txt
aiogram~=2.20
asyncpg~=0.25.0
python-dotenv~=0.20.0
APScheduler~=3.9.1